requirements.txt
pygame
google-generativeai
